requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.42.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.11.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.10
wincertstore==0.2
wrapt==1.11.2
